Update SQL Migrator
